reference-app/backend/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
Flask-OpenTracing==1.1.0
opentracing==2.4.0
jaeger-client==4.8.0
prometheus-client
flask_cors
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-jaeger
prometheus-flask-exporter